setup projects

+++ backend/src/task/task.controller.ts
@@ -10,6 +10,7 @@ import {
 import { TaskService } from './task.service';
 import { CreateTaskDto } from './dto/create-task.dto';
 import { UpdateTaskDto } from './dto/update-task.dto';
+import { Types } from 'mongoose';
 
 @Controller('task')
 export class TaskController {
@@ -27,16 +28,25 @@ export class TaskController {
 
   @Get(':id')
   findOne(@Param('id') id: string) {
+    if (!Types.ObjectId.isValid(id)) {
+      throw new Error('Invalid ObjectId');
+    }
     return this.taskService.findOne(id);
   }
 
   @Patch(':id')
   update(@Param('id') id: string, @Body() updateTaskDto: UpdateTaskDto) {
+    if (!Types.ObjectId.isValid(id)) {
+      throw new Error('Invalid ObjectId');
+    }
     return this.taskService.update(id, updateTaskDto);
   }
 
   @Delete(':id')
   remove(@Param('id') id: string) {
+    if (!Types.ObjectId.isValid(id)) {
+      throw new Error('Invalid ObjectId');
+    }
     return this.taskService.remove(id);
   }
 }
